Add memory_struct definition for arm.
