+++ kernel/kernel_config/src/armv7em/mod.rs
@@ -1,0 +1,7 @@
+cfg_if::cfg_if!{
+
+if #[cfg(target_arch="arm")] {
+    pub mod memory;
+}
+
+}
