+++ kernel/kernel_config/src/lib.rs
@@ -1,4 +1,16 @@
 #![no_std]
 
-pub mod memory;
-pub mod time;
+extern crate cfg_if;
+
+cfg_if::cfg_if!{
+
+if #[cfg(target_arch="x86_64")] {
+    pub mod x86_64;
+    pub use x86_64::memory;
+    pub use x86_64::time;
+} else if #[cfg(target_arch="arm")] {
+    pub mod armv7em;
+    pub use armv7em::memory;
+}
+
+}
