+++ kernel/kernel_config/src/x86_64/mod.rs
@@ -1,0 +1,8 @@
+cfg_if::cfg_if!{
+
+if #[cfg(target_arch="x86_64")] {
+    pub mod memory;
+    pub mod time;
+}
+
+}
