+++ kernel/memory_structs/Cargo.toml
@@ -6,11 +6,8 @@ version = "0.1.0"
 build = "../../build.rs"
 
 [dependencies]
-multiboot2 = "0.7.1"
-# x86_64 = { git = "https://github.com/kevinaboos/x86_64" }
-xmas-elf = { version = "0.6.2", git = "https://github.com/kevinaboos/xmas-elf.git" }
-bit_field = "0.7.0"
 zerocopy = "0.3.0"
+cfg-if = "1.0.0"
 
 [dependencies.derive_more]
 version = "0.14.0"
@@ -20,8 +17,11 @@ features = [ "no_std" ]
 [dependencies.kernel_config]
 path = "../kernel_config"
 
-[dependencies.entryflags_x86_64]
+[target.'cfg(target_arch = "x86_64")'.dependencies.entryflags_x86_64]
 path = "../entryflags_x86_64"
+
+[target.'cfg(target_arch = "x86_64")'.dependencies.bit_field]
+version = "0.7.0"
 
 [lib]
 crate-type = ["rlib"]
